pyproject.toml
[project]
name = "telegram-bot-templates"
version = "0.1.0"
description = ""
authors = [
    {name = "nikewalce",email = "nikewalce1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "telegrinder (>=0.5.1,<0.6.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-telegram-bot (>=22.5,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
